Peer Assignment: Build a Circuit
Build a circuit that contains two pushbuttons, an LED, and any other basic components you think you need. The LED should turn on when either the first button or the second button is pressed. Only use your Arduino for power and ground.

Show the behavior of the system under 3 conditions:
1- no buttons pressed.
2- button 1 pressed alone.
3- button 2 pressed alone.